requirement.txt
torch>=1.7.0
torchvision>=0.8.0
tensorboardX>=2.0
wandb>=0.12.0
numpy>=1.19.0
Pillow>=8.0.0
scipy>=1.5.0
timm>=0.4.12
lmdb>=1.2.0
yaml>=0.2.5
matplotlib>=3.3.0
einops>=0.3.0

